@@ -7,7 +7,7 @@ name = "scos-tekrsa"
 dynamic = ["version"]
 description = "Plugin for SCOS Sensor which adds support for Tektronix RSA-series spectrum analyzers"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = { file = "LICENSE.md" }
 
 authors = [
@@ -32,12 +32,10 @@ classifiers = [
     "Environment :: Plugins",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.13",
 ]
 
 dependencies = [
@@ -48,14 +46,14 @@ dependencies = [
 
 [project.optional-dependencies]
 tests = [
-    "pytest>=7.3.1,<9.0",
-    "pytest-cov>=5.0.0,<6.0",  # pytest-cov >= 6.0 drops python 3.8 support
+    "pytest>=8.0,<9.0",
+    "pytest-cov>=6.0,<7.0",
     "tox>=4.5.1,<5.0",  # Should keep in sync with tox.ini minimum_version
 ]
 
 dev = [
     "hatchling>=1.14.2,<2.0",
-    "pre-commit>=3.3.1,<5.0",
+    "pre-commit>=4.0,<5.0",
     "scos-tekrsa[tests]",
 ]
 
