@@ -1,1 +1,1 @@
-__version__ = "7.0.1"
+__version__ = "8.0.0"
